+++ app/Http/Controllers/CategoryController.php
@@ -5,12 +5,13 @@ namespace App\Http\Controllers;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
     public function index()
     {
-        $categories = Category::all();
+        $categories = Category::where('user_id', Auth::id())->get();
         return Inertia::render('categories/index', [
             'categories' => $categories,
         ]);
@@ -26,12 +27,18 @@ class CategoryController extends Controller
         $validated = $request->validate([
             'name' => 'required',
         ]);
-        $category = Category::create($validated);
+        Category::create([
+            ...$validated,
+            'user_id' => Auth::id(),
+        ]);
         return to_route('categories.index');
     }
 
     public function edit(Category $category)
     {
+        if ($category->user_id !== Auth::id()) {
+            abort(403, 'Unauthorize action.');
+        }
         // dd($request->all());
         return Inertia::render('categories/edit', [
             'category' => $category,
@@ -40,6 +47,9 @@ class CategoryController extends Controller
 
     public function update(Request $request, Category $category)
     {
+        if ($category->user_id !== Auth::id()) {
+            abort(403, 'Unauthorize action.');
+        }
         $validated = $request->validate([
             'name' => 'required',
         ]);
@@ -50,6 +60,9 @@ class CategoryController extends Controller
 
     public function destroy(Category $category)
     {
+        if ($category->user_id !== Auth::id()) {
+            abort(403, 'Unauthorize action.');
+        }
         $category->delete();
         return to_route('categories.index');
     }
